Refactor dependency injection in AdapterModule
The dependency injection in AdapterModule is updated and simplified by allowing Koin to handle the retrieval of the Logger instead of passing it manually, which increases maintainability and consistency. This change is reflected in the updates within adapterModule() method and the frameworkModule in WGSpleefPlugin classes. A test for the plugin was also added to ensure its entry points are working as expected. Furthermore, changes in build.gradle.kts were made for adding testing dependencies and adjusting paths.